% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ALS_CHM_CAU_2014}
\alias{ALS_CHM_CAU_2014}
\title{ALS-derived CHM - Fazenda Cauaxi 2014}
\format{
A \code{SpatRaster} object
}
\source{
ALS data were acquired with support from USAID and the US Department of State with the technical assistance of the Brazilian Corporation
for Agricultural Research (EMBRAPA) and the US Forest Service Office of International Programs via the Sustainable Landscapes Brazil
Project (Keller, M. 2018).
}
\description{
The Airborne Laser Scanning (ALS)-derived Canopy Height Model - (CHM) provided as an example dataset collected in 2014
at Fazenda Cauaxi in the Paragominas Municipality of Pará State,
Brazil, in the eastern Amazon.
}
\details{
The 1-m ALS-CHM was generated using Lastools software (Isenburg 2016)
}
\examples{
library(terra)

ALS_CHM_CAU_2014 <- rast(system.file("tif/ALS_CHM_CAU_2014.tif", package = "ForestGapR"))
plot(ALS_CHM_CAU_2014)
}
\references{
Keller, M. 2018. Available online: https://www.paisagenslidar.cnptia.embrapa.br/webgis/

Isenburg, M. LAStools—Efficient Tools for Lidar Processing. 2018. Available online: http://www.cs.unc.edu/~isenburg/lastools/ (accessed on 3 October 2018).
}
\keyword{ALS_CHM_CAU_2014}
