% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GapsChangesDec.R
\name{GapChangeDec}
\alias{GapChangeDec}
\title{Forest Gaps Changes Detection}
\usage{
GapChangeDec(gap_layer1,gap_layer2)
}
\arguments{
\item{gap_layer1}{ALS-derived gap as an RasterLayer (\code{\link[raster:raster]{raster}}) object at time 1. (output of  \code{\link[GapForestR:getForestGaps]{getForestGaps}} function).}

\item{gap_layer2}{ALS-derived gap as an RasterLayer (\code{\link[raster:raster]{raster}}) object at time 2. (output of  \code{\link[GapForestR:getForestGaps]{getForestGaps}} function).}
}
\value{
A RasterLayer (\code{\link[raster]{raster}}) object representing forest gap change area
}
\description{
This function detects forest canopy gaps changes accross two forest gaps RasterLayer (\code{\link[raster]{raster}}) objects
}
\examples{
\dontrun{
#Loading raster library
library(raster)

# ALS-derived CHM from Fazenda Cauxi - Brazilian tropical forest
data(ALS_CHM_CAU_2012)
data(ALS_CHM_CAU_2014)

# set height tresholds (e.g. 10 meters)
threshold<-10
size<-c(1,1000) # m2

# Detecting forest gaps
gaps_cau2012<-getForestGaps(chm_layer=ALS_CHM_CAU_2012, threshold=threshold, size=size)
gaps_cau2014<-getForestGaps(chm_layer=ALS_CHM_CAU_2014, threshold=threshold, size=size)

# Detecting forest gaps changes
Gap_changes<-GapChangeDec(gap_layer1=gaps_cau2012,gap_layer2=gaps_cau2014)

# Plotting ALS-derived CHM and forest gaps
library(viridis)
par(mfrow=c(1,3))
plot(ALS_CHM_CAU_2012, main="Forest Canopy Gap - 2012", col=viridis(10))
plot(gaps_cau2012, add=TRUE, col="red", legend=FALSE)

plot(ALS_CHM_CAU_2014,  main="Forest Canopy Gap - 2014", col=viridis(10))
plot(gaps_cau2014, add=TRUE,col="blue", legend=FALSE)

plot(ALS_CHM_CAU_2014,main="Forest Gaps Changes Detection",col=viridis(10))
plot(Gap_changes, add=TRUE, col="yellow", legend=FALSE)
}
}
\author{
Carlos Alberto Silva.
}
