% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GapSPDF.R
\name{GapSPDF}
\alias{GapSPDF}
\title{Forest Canopy Gaps as Spatial Polygons}
\usage{
GapSPDF(gap_layer)
}
\arguments{
\item{gap_layer}{ALS-derived gap layer (output of \code{\link[=getForestGaps]{getForestGaps()}} function).
An object of the class SpatRaster.}
}
\value{
A \code{\link[sp:SpatialPointsDataFrame-class]{sp::SpatialPointsDataFrame}} object of the forest canopy gaps.
The result can be converted to a SpatVector and exported as a ESRI shapefile using
\code{\link[terra:writeVector]{terra::writeVector()}} function in the \emph{terra} package.
}
\description{
This function converts forest canopy gaps as \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} to
\code{\link[sp:SpatialPointsDataFrame-class]{sp::SpatialPointsDataFrame}} objects
}
\examples{
# Loading libraries
library(terra)
library(viridis)
library(sf)
library(sp)

# ALS-derived CHM over Adolpho Ducke Forest Reserve - Brazilian tropical forest
ALS_CHM_DUC <- rast(system.file("tif/ALS_CHM_DUC.tif", package = "ForestGapR"))

# set height thresholds (e.g. 10 meters)
threshold <- 10
size <- c(1, 10^4) # m2

# Detecting forest gaps
gaps_duc <- getForestGaps(chm_layer = ALS_CHM_DUC, threshold = threshold, size = size)

# Converting SpatRaster layer to SpatialPolygonsDataFrame
gaps_spdf <- GapSPDF(gap_layer = gaps_duc)

# Plotting ALS-derived CHM and forest gaps
plot(ALS_CHM_DUC, col = viridis(10))
plot(gaps_spdf, add = TRUE, border = "red", lwd = 2)

# Populating the attribute table of Gaps_spdf with gaps statistics
gaps_stats <- GapStats(gap_layer = gaps_duc, chm_layer = ALS_CHM_DUC)
gaps_spdf <- merge(gaps_spdf, gaps_stats, by = "gap_id")
head(gaps_spdf@data)
# Convert to SpatVector for export 
gaps_vect <- terra::vect(gaps_spdf)
# Save the SpatVector object to the Shapefile
\dontrun{
terra::writeVector(gaps_vect, 'output_shapefile.shp', overwrite = TRUE)
}
}
\author{
Carlos Alberto Silva and Lucy Beese.
}
