\name{ALS_CHM_DUC}
\alias{ALS_CHM_DUC}
\docType{data}
\title{ALS-derived CHM - Adolpho Ducke Forest Reserve 2012}
\description{
The Airborne Laser Scanning (ALS)-derived Canopy Height Model - (CHM) provided as an example dataset collected in 2012 
at the Adolpho Ducke Forest Reserve in Municipality of Manaus of State of Amazonas,
Brazil, in central Amazon. 
}
\usage{data(ALS_CHM_DUC)}
\format{
  The format is:'RasterLayer'
}
\details{
The 1-m ALS-CHM was generated using Lastools software (Isenburg 2016) 
}
\source{
ALS data were acquired with support from USAID and the US Department of State with the technical assistance of the Brazilian Corporation
for Agricultural Research (EMBRAPA) and the US Forest Service Office of International Programs via the Sustainable Landscapes Brazil 
Project (Keller, M. 2018). 
}
\references{
Keller, M. 2018. Available online: https://www.paisagenslidar.cnptia.embrapa.br/webgis/

Isenburg, M. LAStools—Efficient Tools for Lidar Processing. 2018. Available online: http://www.cs.unc.
edu/~isenburg/lastools/ (accessed on 3 October 2018).
}
\examples{
data(ALS_CHM_DUC)
## plot(ALS_CHM_DUC)
}
\keyword{ALS_CHM_DUC}